Migrate to esm
In this diff I convert source to esm so we can easily use this lib with
native browser modules. This also allow bundlers do not add additional
runtime to work around cjs.

I renamed index.js to index.esm.js and conveted it to esm. Enabled in
tests esm loader which just works. For node users converted to cjs with
rollup.